requirements.txt
flask==3.0.0
requests==2.31.0
pymysql==1.1.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
schedule==1.2.1
pydantic~=2.5.3
mypy==1.8.0
Werkzeug~=3.0.1

# In case of using LDAP
ldap3==2.9.1
types-ldap3==2.9.13.20240106

